Smaller types to make smaller position

+++ pkg/bitboard/bitscan.go
@@ -2,16 +2,16 @@ package bitboard
 
 import "math/bits"
 
-func LeastSignificantOneBit(b Bitboard) int {
+func LeastSignificantOneBit(b Bitboard) uint8 {
 	if b == Empty {
 		panic("Bitboard empty")
 	}
-	return bits.TrailingZeros64(uint64(b))
+	return uint8(bits.TrailingZeros64(uint64(b)))
 }
 
-func MostSignificantOneBit(b Bitboard) int {
+func MostSignificantOneBit(b Bitboard) uint8 {
 	if b == Empty {
 		panic("Bitboard empty")
 	}
-	return bits.LeadingZeros64(uint64(b))
+	return uint8(bits.LeadingZeros64(uint64(b)))
 }
