+++ pkg/pieces/queen/queen.go
@@ -7,6 +7,6 @@ import (
 )
 
 // AttacksBySquare returns the attacks for a given square.
-func AttacksBySquare(square int, occupied bitboard.Bitboard) bitboard.Bitboard {
+func AttacksBySquare(square uint8, occupied bitboard.Bitboard) bitboard.Bitboard {
 	return rook.AttacksBySquare(square, occupied) | bishop.AttacksBySquare(square, occupied)
 }
